add net7.0 and add all files to project
