+++ PSModules/PSModules/CounterElement.xaml.cs
@@ -8,73 +8,73 @@ using System.Windows.Controls;
 using System.Windows.Data;
 using System.Windows.Documents;
 using System.Windows.Input;
-using System.Windows.Media; 
+using System.Windows.Media;
 using System.Windows.Media.Imaging;
 using System.Windows.Navigation;
 using System.Windows.Shapes;
 
-namespace clown_mega_project
+namespace PSModules
 {
     /// <summary>
-    /// Логика взаимодействия для Random.xaml
+    /// Логика взаимодействия для CounterElement.xaml
     /// </summary>
-    public partial class RandomGenerator : UserControl
+    public partial class CounterElement : UserControl
     {
-        private readonly Random rand = new Random();
-        private int last_value = 0;
+        public int value = 0;
 
         double btnFontSize = 6;         // процент от высоты окна
         double labelFontSize = 12;
 
-        private void Element_Loaded(object sender, RoutedEventArgs e)
-        {
-            double labelfontsize = Application.Current.MainWindow.Height * (labelFontSize / 100);
-            double btnfontsize = App.Current.MainWindow.Height * (btnFontSize / 100);
-            System.Windows.Application.Current.Resources.Remove("LabelFontSize");
-            System.Windows.Application.Current.Resources.Add("LabelFontSize", labelfontsize);
-            System.Windows.Application.Current.Resources.Remove("BtnFontSize");
-            System.Windows.Application.Current.Resources.Add("BtnFontSize", btnfontsize);
-        }
-
-        private void Element_Resized(object sender, SizeChangedEventArgs e)
-        {
-            double labelfontsize = Application.Current.MainWindow.Height * (labelFontSize / 100);
-            double btnfontsize = App.Current.MainWindow.Height * (btnFontSize / 100);
-            System.Windows.Application.Current.Resources.Remove("LabelFontSize");
-            System.Windows.Application.Current.Resources.Add("LabelFontSize", labelfontsize);
-            System.Windows.Application.Current.Resources.Remove("BtnFontSize");
-            System.Windows.Application.Current.Resources.Add("BtnFontSize", btnfontsize);
-        }
-
-        public RandomGenerator()
+        public CounterElement()
         {
             InitializeComponent();
         }
 
-        private void Generate_btn_Click(object sender, RoutedEventArgs e)
-        {
-            int from, to;
-
-            try
-            {
-                from = Convert.ToInt32(From_textbox.Text);
-                to = Convert.ToInt32(To_textbox.Text);
-                Result_textblock.Text = rand.Next(from, to).ToString();
-            }
-            catch {
-                Result_textblock.Text = "Not a number";
-            }
-
+        private void Element_Loaded(object sender, RoutedEventArgs e)
+        {/*
+            double labelfontsize = Application.Current.MainWindow.Height * (labelFontSize / 100);
+            double btnfontsize = App.Current.MainWindow.Height * (btnFontSize / 100);
+            System.Windows.Application.Current.Resources.Remove("LabelFontSize");
+            System.Windows.Application.Current.Resources.Add("LabelFontSize", labelfontsize);
+            System.Windows.Application.Current.Resources.Remove("BtnFontSize");
+            System.Windows.Application.Current.Resources.Add("BtnFontSize", btnfontsize);*/
         }
 
-        private void Set_Value(int _value)
-        {
-            Result_textblock.Text = _value.ToString();
-            last_value = _value;
+        private void Element_Resized(object sender, SizeChangedEventArgs e)
+        {/*
+            double labelfontsize = Application.Current.MainWindow.Height * (labelFontSize / 100);
+            double btnfontsize = App.Current.MainWindow.Height * (btnFontSize / 100);
+            System.Windows.Application.Current.Resources.Remove("LabelFontSize");
+            System.Windows.Application.Current.Resources.Add("LabelFontSize", labelfontsize);
+            System.Windows.Application.Current.Resources.Remove("BtnFontSize");
+            System.Windows.Application.Current.Resources.Add("BtnFontSize", btnfontsize);*/
         }
 
-        public int GetValue() => last_value;
+        private void Minus_btn_Click(object sender, RoutedEventArgs e)
+        {
+            value--;
+            Update_text();
+        }
 
-        public void SetValue(int _value) => Set_Value(_value);
+        private void Reset_btn_Click(object sender, RoutedEventArgs e)
+        {
+            value = 0;
+            Update_text();
+        }
+
+        private void Plus_btn_Click(object sender, RoutedEventArgs e)
+        {
+            value++;
+            Update_text();
+        }
+
+        private void Update_text()
+        {
+            Value_label.Content = value.ToString();
+        }
+
+        public int GetValue() => value;
+
+        public void SetValue(int _value) => value = _value;
     }
 }
