+++ PSModules/PSModules/TimerElement.xaml.cs
@@ -16,7 +16,7 @@ using System.Windows.Navigation;
 using System.Windows.Shapes;
 using System.Windows.Threading;
 
-namespace clown_mega_project
+namespace PSModules
 {
     /// <summary>
     /// Логика взаимодействия для TimerElement.xaml
@@ -34,23 +34,23 @@ namespace clown_mega_project
         double labelFontSize = 12;
 
         private void Element_Loaded(object sender, RoutedEventArgs e)
-        {
+        {/*
             double labelfontsize = Application.Current.MainWindow.Height * (labelFontSize / 100);
             double btnfontsize = App.Current.MainWindow.Height * (btnFontSize / 100);
             System.Windows.Application.Current.Resources.Remove("LabelFontSize");
             System.Windows.Application.Current.Resources.Add("LabelFontSize", labelfontsize);
             System.Windows.Application.Current.Resources.Remove("BtnFontSize");
-            System.Windows.Application.Current.Resources.Add("BtnFontSize", btnfontsize);
+            System.Windows.Application.Current.Resources.Add("BtnFontSize", btnfontsize);*/
         }
 
         private void Element_Resized(object sender, SizeChangedEventArgs e)
-        {
+        {/*
             double labelfontsize = Application.Current.MainWindow.Height * (labelFontSize / 100);
             double btnfontsize = App.Current.MainWindow.Height * (btnFontSize / 100);
             System.Windows.Application.Current.Resources.Remove("LabelFontSize");
             System.Windows.Application.Current.Resources.Add("LabelFontSize", labelfontsize);
             System.Windows.Application.Current.Resources.Remove("BtnFontSize");
-            System.Windows.Application.Current.Resources.Add("BtnFontSize", btnfontsize);
+            System.Windows.Application.Current.Resources.Add("BtnFontSize", btnfontsize);*/
         }
 
         public TimerElement()
