+++ PSModules/PSModules/ToggleSwitch.xaml
@@ -1,9 +1,9 @@
-﻿<UserControl x:Class="clown_mega_project.ToggleSwitch"
+﻿<UserControl x:Class="PSModules.ToggleSwitch"
              xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
              xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
              xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006" 
              xmlns:d="http://schemas.microsoft.com/expression/blend/2008" 
-             xmlns:local="clr-namespace:clown_mega_project"
+             xmlns:local="clr-namespace:PSModules"
              mc:Ignorable="d" 
              d:DesignHeight="100" d:DesignWidth="200" 
              Name="ToggleUserControl">
