+++ PSModules/PSModules/dice_d6.xaml
@@ -1,9 +1,9 @@
-﻿<UserControl x:Class="clown_mega_project.dice_d6"
+﻿<UserControl x:Class="PSModules.dice_d6"
              xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
              xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
              xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006" 
              xmlns:d="http://schemas.microsoft.com/expression/blend/2008" 
-             xmlns:local="clr-namespace:clown_mega_project"
+             xmlns:local="clr-namespace:PSModules"
              mc:Ignorable="d" 
              Loaded="Element_Loaded"
              SizeChanged="Element_Resized"
