+++ PlayAssistant/App.xaml.cs
@@ -9,7 +9,7 @@ using System.Windows;
 namespace PlayAssistant
 {
     /// <summary>
-    /// Логика взаимодействия для App.xaml
+    /// Interaction logic for App.xaml
     /// </summary>
     public partial class App : Application
     {
