+++ PlayAssistant/GameChooseMenu.xaml
@@ -1,9 +1,9 @@
-﻿<Page x:Class="clown_mega_project.GameChooseMenu"
+﻿<Page x:Class="PlayAssistant.GameChooseMenu"
       xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
       xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
       xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006" 
       xmlns:d="http://schemas.microsoft.com/expression/blend/2008" 
-      xmlns:local="clr-namespace:clown_mega_project"
+      xmlns:local="clr-namespace:PlayAssistant"
       mc:Ignorable="d" 
       x:Name="GameChooseMenu_element"
       d:DesignHeight="450" d:DesignWidth="800" MinHeight="450" MinWidth="800"
