+++ PlayAssistant/TestGamePage.xaml
@@ -1,9 +1,9 @@
-﻿<Page x:Class="clown_mega_project.TestGamePage"
+﻿<Page x:Class="PlayAssistant.TestGamePage"
       xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
       xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
       xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006" 
       xmlns:d="http://schemas.microsoft.com/expression/blend/2008" 
-      xmlns:local="clr-namespace:clown_mega_project"
+      xmlns:local="clr-namespace:PlayAssistant"
       mc:Ignorable="d" 
       d:DesignHeight="450" d:DesignWidth="800"
       Title="TestGamePage">
